Se hizo la migración a TypeScripts, ahora estamos usando react-typescript y tailwind
