+++ gimnasioReact/.gitignore
@@ -12,11 +12,6 @@ dist
 dist-ssr
 *.local
 
-.env
-.env.local
-.env.development
-.env.production
-
 # Editor directories and files
 .vscode/*
 !.vscode/extensions.json
