+++ gimnasioReact/src/components/Table.tsx
