+++ gimnasioReact/src/index.css
@@ -28,5 +28,3 @@ tbody tr:nth-child(even) {
     background-color: #2928287c;
     
 }
-
-
