+++ gimnasioReact/src/pages/admin/registroPorDia/MemberFormDay.tsx
@@ -4,14 +4,14 @@ import { useParams, useNavigate } from "react-router-dom";
 import { useEffect } from "react";
 //Icons
 import { CiUser } from 'react-icons/ci';
-import { RiLockPasswordLine, RiLoginBoxLine } from "react-icons/ri";
+import { RiLoginBoxLine } from "react-icons/ri";
 import { MdOutlinePhoneAndroid, MdOutlinePriceChange } from "react-icons/md";
 import { BsCalendar2Date } from "react-icons/bs";
 //Mensajes
 import { toast } from "react-hot-toast";
 
 //ui
-import { Input, Label, Button } from '../../../component/ui/index';
+import { Input, Label, Button } from '../../../components/ui/index';
 
 //API
 import { createMemberDay, getMember, updateMember } from '../../../api/userGymDay.api';
@@ -34,7 +34,7 @@ const RegisterMiembroDay = () => {
                         background: '#4b5563',   // Fondo negro
                         color: '#fff',           // Texto blanco
                         padding: '16px',
-                        borderRadios: '8px',
+                        borderRadius: '8px',
                     },
 
                 });
