+++ gimnasioReact/src/pages/auth/Register.tsx
@@ -6,7 +6,7 @@ import { RiLoginBoxLine, RiMailFill, RiLockFill,  RiUserLine } from "react-icons
 import { toast } from 'react-hot-toast';
 
 //ui
-import { Input, Label, Button } from '../../component/ui/index';
+import { Input, Label, Button } from '../../components/ui/index';
 
 //api
 import { CreateUsers } from "../../api/users.api";
